Extract `QuizGameEngine` interfaces to a protocol `QuizGame`
